Fixed aura time calculation overflow
I think what happened is, AURA_CALCTIME overflowed during calculation when Player.Aura.Time reached a high enough value.

Tested auras up to an hour without issue.

+++ DoomRPG/scripts/Skills.c
@@ -1372,9 +1372,9 @@ NamedScript Console bool UseAura(SkillLevelInfo *SkillLevel, void *Data)
 
     // Should the timer be stacked because you used the same Aura?
     if (Stack || Player.Perks[STAT_ENERGY])
-        Player.Aura.Time += AURA_CALCTIME;
+        Player.Aura.Time += (long int)AURA_CALCTIME;
     else
-        Player.Aura.Time = AURA_CALCTIME;
+        Player.Aura.Time = (long int)AURA_CALCTIME;
 
     // Apply Aura
     Player.Aura.Type[Index].Active = true;
